+++ packages/demo-app/src/createInitialState.js
@@ -1,1 +1,4 @@
-export default () => JSON.parse(document.getElementById('root').getAttribute('data-initialState'));
+export default () => ({ page: { title: 'Modular Toolkit Demo' } });
+
+// Note: you can also use this module to hydrate from an existing state (e.g. from server side rendering) like this:
+// export default () => JSON.parse(document.getElementById('root').getAttribute('data-initialState'));