+++ packages/demo-module/src/components/HackerNews/HackerNewsContainer.js
@@ -1,9 +1,10 @@
-import { connectSelectors } from '@modular-toolkit/hocs';
+import { connectSelectors, bootstrap } from '@modular-toolkit/hocs';
 import { compose, withHandlers, setDisplayName } from 'recompose';
 
 import { enhanceWithSpinner } from '../../enhancers';
 import { handleLoadingTopStories } from '../../handlers';
 import { selectTopStories } from '../../selectors';
+import { loadTopStoriesAction } from '../../actions';
 import HackerNews from './HackerNews';
 
 const enhance = compose(
@@ -11,6 +12,7 @@ const enhance = compose(
     connectSelectors({
         topStories: selectTopStories
     }),
+    bootstrap(({ dispatch, topStories }) => !topStories && dispatch(loadTopStoriesAction.start())),
     withHandlers({ handleLoadingTopStories }),
 
     enhanceWithSpinner