+++ packages/hocs/src/bootstrap.test.js
@@ -1,7 +1,6 @@
 import bootstrap from './bootstrap';
-import { spy } from 'sinon';
 
-const mockLifecycle = spy();
+const mockLifecycle = jest.fn();
 
 jest.mock('recompose', () => ({
     lifecycle: (...args) => mockLifecycle(...args)
@@ -9,7 +8,11 @@ jest.mock('recompose', () => ({
 
 describe('When I call the bootstrap function with some callback function', () => {
     beforeEach(() => bootstrap(() => {}));
-    describe('the lifecycle method of recompose', () => {
-        it('is called', () => void mockLifecycle.should.have.been.called);
-    });
+    describe('the lifecycle method of recompose', () =>
+        it('is called', () =>
+            expect(mockLifecycle).toBeCalledWith(
+                expect.objectContaining({
+                    componentDidMount: expect.any(Function)
+                })
+            )));
 });