+++ packages/demo-app/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 6.0.0 / 17 Oct 2018
+
+* Hacker News data loaded after page load instead of hydrating previous state
+* Updated to create-react-app v2
+* Dependency update, now using demo-module 6.0.0
+
 ## 5.0.2 / 16 Oct 2018
 
 * Dependency update, now using demo-module 5.0.1